Fix 1550043 and 1548314
git-svn-id: http://svn.terracotta.org/svn/ehcache@186 b9324663-ca0f-0410-8574-be9b3887307d

+++ trunk/src/main/java/net/sf/ehcache/Element.java
@@ -48,7 +48,7 @@ public final class Element implements Serializable, Cloneable {
 
     private static final Log LOG = LogFactory.getLog(Element.class.getName());
 
-    private static final int ONE_SECOND = 1000;
+    private static final long ONE_SECOND = 1000L;
 
     /**
      * the cache key.
